Fixed bugs in delete news comment functionality
Resolves #36 and resolves #37

+++ nieuwsDelete.php
@@ -19,52 +19,56 @@ $cTPL = new Template('Templates/main.tpl');
 connectDB();
 
 /* Permissie controleren */
-if (($cUser->checkSession()) || ($cUser->checkCookie())) {
+if (($cUser->checkSession() || $cUser->checkCookie()) && isset($_GET['id'])) {
     $sQuery = "SELECT userid FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
     if ($cResult = mysql_query($sQuery)) {
         $aData = mysql_fetch_assoc($cResult);
-        if (($cUser->m_iPermis & 2) || ($cUser->m_iUserid == $aData['userid'])) {
-            /* Controleren of het formulier verzonden is */
-            if (isset($_POST['delete'])) {
-                $sQuery = "SELECT nieuwsid FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
-                if ($cResult = mysql_query($sQuery)) {
-                    $sQuery = "DELETE FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
-                    if (mysql_query($sQuery)) {
-                        $aData = mysql_fetch_assoc($cResult);
-                        header('Location: shownieuws.php?id=' . $aData['nieuwsid']);
+        if($aData) {
+            if (($cUser->m_iPermis & 2) || ($cUser->m_iUserid == $aData['userid'])) {
+                /* Controleren of het formulier verzonden is */
+                if (isset($_POST['delete'])) {
+                    $sQuery = "SELECT nieuwsid FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
+                    if ($cResult = mysql_query($sQuery)) {
+                        $sQuery = "DELETE FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
+                        if (mysql_query($sQuery)) {
+                            $aData = mysql_fetch_assoc($cResult);
+                            header('Location: shownieuws.php?id=' . $aData['nieuwsid']);
+                        } else {
+                            $cTPL->setPlace('TITEL', 'Fout met database');
+                            $cTPL->setPlace('CONTENT', 'Er is iets fout gegaan met de database');
+                            $cTPL->show();
+                        }
                     } else {
                         $cTPL->setPlace('TITEL', 'Fout met database');
                         $cTPL->setPlace('CONTENT', 'Er is iets fout gegaan met de database');
                         $cTPL->show();
                     }
                 } else {
-                    $cTPL->setPlace('TITEL', 'Fout met database');
-                    $cTPL->setPlace('CONTENT', 'Er is iets fout gegaan met de database');
+                    $cTPL->setPlace('TITEL', 'Reactie verwijderen');
+                    $cTPL->setBlock('LOGIN', 'logout');
+                    $cTPL->parse();
+
+                    if ($cUser->m_iPermis & 2) {
+                        $cTPL->setBlock('ADMIN', 'admin');
+                    }
+
+                    $cTPL->setFile('CONTENT', 'Templates/nieuwsDelete.tpl');
+                    $cTPL->parse();
+
+                    $sQuery = "SELECT nieuwsid FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
+                    if ($cResult = mysql_query($sQuery)) {
+                        $aData = mysql_fetch_assoc($cResult);
+                        $cTPL->setPlace('NIEUWSID', $aData['nieuwsid']);
+                    }
+                    $cTPL->setPlace('ID', $_GET['id']);
+
                     $cTPL->show();
                 }
             } else {
-                $cTPL->setPlace('TITEL', 'Reactie verwijderen');
-                $cTPL->setBlock('LOGIN', 'logout');
-                $cTPL->parse();
-
-                if ($cUser->m_iPermis & 2) {
-                    $cTPL->setBlock('ADMIN', 'admin');
-                }
-
-                $cTPL->setFile('CONTENT', 'Templates/nieuwsDelete.tpl');
-                $cTPL->parse();
-
-                $sQuery = "SELECT nieuwsid FROM nieuwsreacties WHERE reactieid='" . add($_GET['id']) . "';";
-                if ($cResult = mysql_query($sQuery)) {
-                    $aData = mysql_fetch_assoc($cResult);
-                    $cTPL->setPlace('NIEUWSID', $aData['nieuwsid']);
-                }
-                $cTPL->setPlace('ID', $_GET['id']);
-
-                $cTPL->show();
+                header('HTTP/1.0 404');
             }
         } else {
-            header('HTTP/1.0 404');
+            header('hTTP/1.0 404');
         }
     } else {
         print 'Er is iets fout gegaan met de mysql database';
